Format and move types to separate file
